requirements.txt
gitpython
regex